Refactored Reducer file into discrete models for each component of the state
