+++ frontend/src/js/models/BaseModel.js
@@ -1,0 +1,11 @@
+class BaseModel {
+    static reduce(state = {}, action) {
+        return state;
+    }
+
+    static getInitialState() {
+        return Promise.resolve({});
+    }
+}
+
+export default BaseModel
